backend/requirements.txt
# ---------------------------
# Web Scraping
# ---------------------------
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=5.1.0
trafilatura>=1.7.0
webdriver-manager>=4.0.1
selenium>=4.18.0

# ---------------------------
# NLP (core)
# ---------------------------
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.11.0

# Sentence embeddings
sentence-transformers>=2.6.0

# ---------------------------
# Transformers / Models
# ---------------------------
torch>=2.2.0
transformers==4.38.2
tokenizers==0.15.2
accelerate==0.27.2
google-generativeai>=0.5.0
python-dotenv>=1.0.0

sentencepiece>=0.2.0
protobuf>=4.25.3

# fallback tokenizer (uneori cerut intern)
tiktoken>=0.7.0

# ---------------------------
# Language detection
# ---------------------------
langdetect>=1.0.9

# ---------------------------
# API Server
# ---------------------------
flask>=3.0.0
flask-cors>=4.0.0

# ---------------------------
# Utilities
# ---------------------------
tqdm>=4.66.0
python-dotenv>=1.0.0
huggingface_hub>=0.23.0
